+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # brote
 
+## 1.0.2
+
+### Patch Changes
+
+- c7598de: adding daily progress to home page
+
 ## 1.0.1
 
 ### Patch Changes