+++ src/components/pageHome/CHANGELOG.md
@@ -1,5 +1,11 @@
 # home-page
 
+## 0.0.2
+
+### Patch Changes
+
+- c7598de: adding daily progress to home page
+
 ## 0.0.1
 
 ### Patch Changes